backend/requirements.txt
fastapi
uvicorn
web3
pydantic
numpy
pandas
python-dotenv
requests
scikit-learn